Finished the last challenge for chapter 5.
Added a Cheater boolean and getter and setter on the QuestionModel. Check that boolean every time checkAnswer is called to determine whether or not to show the judgement toast. The code is ugly though and I should clean it up. Not really sure why I'm using mIsCheater in onActivityResult AND calling the getter on checkAnswer. Feel like that can be consolidated.