Fixed a bug where, on orientation change, it was switching mIsCheater to false and not showing the judgement toast. This was caused because...
On coming back from the saved instance state, I had set the mIsCheater boolean in MainActivity to the boolean value pulled from savedinstance state. That's all is fine and dandy. However, I had 2 "if (savedInstanceState != true)" statements at the top and bottom of the onCreate method. The first one set the boolean value, the second set the index. Because the first statement (setting the boolean) created a savedInstance state, the second "if (savedInstanceState != true)", which updated the index from the bundle, was never called. Now my activity does not have the right index/question.

After this, I would call updateQuestion(). In that method, I call mIsCheat = mQuestionBank[mCurrentIndex].isCheat(); - effectively assigning mIsCheat to the boolean value of the wrong question.

I should look into what exactly i'm doing in the updateQuestion() method, because in this scenario I don't need to be assigning mIsCheat to anything, since I've already declared it in onCreate. However, not sure if ripping that statement out of the method is the best solution here

--

That is an old commit note ^

Right now, I'm just trying to figure out how to make the mQuestionBank parcelable and pass it around via savedInstanceState